+++ dataquality/__init__.py
@@ -30,7 +30,7 @@ If you want to train without a model, you can use the auto framework:
         dataquality.get_insights()
 """
 
-__version__ = "v0.8.18a3"
+__version__ = "v0.8.18a4"
 
 import sys
 from typing import Any, List, Optional
