+++ dataquality/loggers/data_logger/text_classification.py
@@ -479,7 +479,7 @@ class TextClassificationDataLogger(BaseGalileoDataLogger):
         else:
             emb_cols = ["id", "emb", "x", "y", "emb_pca"]
             emb_cols = [c for c in emb_cols if c in df_copy.get_column_names()]
-            ignore_cols = ["emb", "split_id"] + prob_cols
+            ignore_cols = ["split_id"] + prob_cols + emb_cols
             other_cols = [i for i in df_copy.get_column_names() if i not in ignore_cols]
             other_cols += ["id"]
