+++ dataquality/loggers/model_logger/base_model_logger.py
@@ -17,7 +17,7 @@ from dataquality.schemas.task_type import TaskType
 from dataquality.utils.ampli import AmpliMetric
 from dataquality.utils.dq_logger import get_dq_logger
 from dataquality.utils.thread_pool import ThreadPoolManager
-from dataquality.utils.vaex import _save_hdf5_file
+from dataquality.utils.hdf5_store import _save_hdf5_file
 
 analytics = Analytics(ApiClient, config)  # type: ignore
